Added readme updated design

+++ controllers/advideo.php
@@ -7,12 +7,12 @@ include("../models/CourseModel.php");
 include_once("Auth.php");
 
 // check if user is $login;
-if(!Auth::isLogin()){
+if (!Auth::isLogin()) {
     header("Location" . "../views/Login.php");
     exit;
 }
 // check if user is admin 
-if(!Auth::isAdminUser()){
+if (!Auth::isAdminUser()) {
     header("Location:" . "../views/partials/404.php");
     exit;
 }
@@ -30,8 +30,7 @@ class AddVideo
             $this->course_id = $_POST['courseId'];
             $this->video_title = $_POST['video-title'];
             // echo json_encode($_POST);   
-        }
-        else{
+        } else {
             header('Location:' . '../views/partials/404.php');
             exit;
         }
@@ -52,55 +51,36 @@ class AddVideo
         if (isset($_FILES["videoFile"])) {
             // hashing the video file 
             $videoname = basename($_FILES['videoFile']['name']);
-            // $hash_value = hash_file('sha1', $_FILES['videoFile']['tmp_name']);
+            $hash_value = hash_file('sha1', $_FILES['videoFile']['tmp_name']);
             //  echo json_encode($hash_value);
 
             $extension = pathinfo($videoname, PATHINFO_EXTENSION);
-            // $videofilename =   "/uploads/videos/" . substr($hash_value, 0, 6) . '.' . $extension;
+            $videofilename =   "/uploads/videos/" . substr($hash_value, 0, 6) . '.' . $extension;
             // echo json_encode($videofilename);
-            
+
             $videoname = strtolower(str_replace(' ', '', basename($_FILES['videoFile']['name'])));
             $current_date_time = date('YmdHis');
             $filename = $videoname . $current_date_time;
             $hashedcode = hash('sha1', $filename);
             $file = substr($hashedcode, 0, 6);
-            $target_file = $target_dir . $file . '.' . $extension;
-
-            // if (!file_exists($videofilename)) {
-            //     move_uploaded_file($_FILES['$videoFile']['tmp_name'],$_SERVER['DOCUMENT_ROOT'] . $videofilename);
-            //     echo json_encode("sj");
-            // }
-            // else{
-            //     echo json_encode("sdcs");
-            // }
-
-            // if (symlink($videofilename, $target_file)) {
-            //     // echo "Symbolic link created successfully.";
-            // } else {
-            //     // echo "Error creating symbolic link.";
-            // }
-            // upload video file in the directory 
-             try {
-                 File::uploadFile($_FILES["videoFile"]["tmp_name"], $_SERVER['DOCUMENT_ROOT'] . $target_file);
-             } catch (Exception $e) {
-                 echo json_encode(array("error" => $e->getMessage()));
-                 exit;
-             }
-            
+            $target_file = $target_dir . $file . '.txt';
+            if (!file_exists($videofilename)) {
+                // upload video file in the directory 
+                try {
+                    File::uploadFile($_FILES["videoFile"]["tmp_name"], $_SERVER['DOCUMENT_ROOT'] . $videofilename);
+                } catch (Exception $e) {
+                    echo json_encode(array("error" => $e->getMessage()));
+                    exit;
+                }
+            } else {
+                echo json_encode("sdcs");
+            }
+            $fileHandle = fopen($target_file, "w") or die("Unable to open file!");
+            fwrite($fileHandle, $videofilename);
+            fclose($fileHandle);
         }
-        
-        // save video details in database 
-        try {
-            $videoobj = new Video();
-            $id = $videoobj->createVideo($this->video_title, $target_file, $this->section_id);
-            echo json_encode(array('title' => $this->video_title, 'id' => $id, "url" => $target_file));
-        } catch (Exception $e) {
-            echo json_encode(array("error" => $e->getMessage()));
-        }
-        
     }
 }
 
 $addobj = new AddVideo();
-
 $addobj->addVideo();
